Version 0.4.0 Fix odd javascript math errors causing very long fractional seconds Previously to get the seconds value we were executing `seconds + (milliseconds / 1000)` where seconds could be 0 - 59 and milliseconds could be 0 - 999. Javascript math can be strange apparently.  For example dividing 804 by 1000 results in 0.804, but performing `3 + (804 / 1000)` results in `3.8040000000000003` :eyeroll: Even separating the math still results in the same problem: ``` let ms = 804 / 1000; let seconds = 3 + ms;  // => 3.8040000000000003 ``` Some obvious ones we saw a lot of are: * `x.8040000000000003` * `x.3469999999999995` * `x.7880000000000003` * `x.5739999999999998`
Avoid math altogether - since we want a maximum of 3 decimal digits anyway, just turn milliseconds into a string, zero-pad it to 3 places (e.g. `0804`) then get seconds as a string, append a dot and the millisecond string; and finally parse it as a float.  Voila` - `3.804`

(also raise the minutes cutoff to 5 so that 0 - 4 seconds contain millisecond fractions.  5 seconds and above will not.